requirements.txt
flask
gunicorn
google-generativeai
wikipedia